# Docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "labgate \u2014 2PP Lab Control Platform",
    "description": "\nDeterministic control platform for the 2PP laser fabrication rig.\n\nLifecycle: draft -> validated -> approved -> queued -> running -> completed.\nA plan is a declarative Experiment Specification; it is validated against\nhard bounds, dry-runnable without hardware, approved by a qualified human\n(never its own proposer), executed by the platform's deterministic engine,\nand fully audited. AI planners are ordinary clients of this API.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/capabilities": {
      "get": {
        "summary": "Get hardware capabilities",
        "description": "Query machine-readable grounding bounds, devices, and operations supported by this rig.",
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Getcapabilities"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/devices": {
      "get": {
        "summary": "Get device states",
        "description": "Get real-time live connection status and coordinates for all active devices.",
        "operationId": "getDevices",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Getdevices"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/plans": {
      "get": {
        "summary": "List experiment plans",
        "description": "List summaries of all submitted experiment plans and their status.",
        "operationId": "listPlans",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanSummary"
                  },
                  "type": "array",
                  "title": "Response Listplans"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Submit experiment plan",
        "description": "Submit a declarative Experiment Specification for validation and approval. Auto-runs validation checks.",
        "operationId": "submitPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/plans/{plan_id}": {
      "get": {
        "summary": "Get plan details",
        "description": "Get full plan record including validation report, state, and identity history.",
        "operationId": "getPlan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getplan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plans/{plan_id}/dry-run": {
      "post": {
        "summary": "Dry run & preview plan",
        "description": "Estimate motion distance, duration, and render toolpath preview PNG for approval.",
        "operationId": "dryRunPlan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Dryrunplan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plans/{plan_id}/rerun": {
      "post": {
        "summary": "Clone & rerun plan",
        "description": "Clone an existing plan's recipe into a fresh plan that starts from draft/validated state.",
        "operationId": "rerunPlan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/queue": {
      "get": {
        "summary": "Get execution queue",
        "description": "View current running plan and queued plans in the execution engine.",
        "operationId": "getQueue",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Getqueue"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/models": {
      "get": {
        "summary": "List STL models",
        "description": "List all uploaded STL 3D models stored on the platform.",
        "operationId": "listModels",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Listmodels"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Upload STL model",
        "description": "Upload a 3D STL mesh model file for print_stl operations.",
        "operationId": "uploadModel",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Uploadmodel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/models/{model_id}": {
      "get": {
        "summary": "Get STL model info",
        "description": "Get metadata, dimensions, and point count for an uploaded STL model.",
        "operationId": "getModelInfo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getmodelinfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plans/{plan_id}/approve": {
      "post": {
        "summary": "Approve experiment plan",
        "description": "Human approver authorizes a plan for execution. Enforces proposer != approver.",
        "operationId": "approvePlan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plans/{plan_id}/execute": {
      "post": {
        "summary": "Execute approved plan",
        "description": "Enqueue an approved plan for execution on hardware or simulator.",
        "operationId": "executePlan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plans/{plan_id}/abort": {
      "post": {
        "summary": "Abort plan execution",
        "description": "Cooperatively abort a queued or executing plan and immediately safe-state hardware.",
        "operationId": "abortPlan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Abortplan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plans/{plan_id}/results": {
      "get": {
        "summary": "Get plan results",
        "description": "Fetch telemetry events and list of generated artifacts for a completed run.",
        "operationId": "getPlanResults",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getplanresults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlanSummary": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "proposer": {
            "type": "string",
            "title": "Proposer"
          },
          "approver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver"
          },
          "validation_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Summary"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "title",
          "state",
          "proposer",
          "approver",
          "validation_summary",
          "created_at"
        ],
        "title": "PlanSummary"
      },
...
      "SubmitPlanRequest": {
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/ExperimentSpec"
          }
        },
        "type": "object",
        "required": [
          "spec"
        ],
        "title": "SubmitPlanRequest"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Bearer token identifying the operator or approver user identity"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8523",
      "description": "Local Labgate API Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}